Move registry bootstrap tests out of prod file
